+++ src/services/entities/board/boardLike.entity.ts
@@ -10,7 +10,7 @@ import { Board } from './board.entity';
 
 @Entity({ name: 'boardLike' })
 export class BoardLike {
-  @PrimaryGeneratedColumn({ type: 'int', name: 'id' })
+  @PrimaryGeneratedColumn({ type: 'bigint', name: 'id' })
   id: number;
 
   @Column({ type: 'tinyint', name: 'likeStatus', default: 0 })
